Moved existing slack code into separate module.
