+++ src/index.js
@@ -1,7 +1,8 @@
-const Slack = require('slack');
+// const Slack = require('slack');
 const Airtable = require('airtable');
 
 const { getCoords, distanceBetweenCoords } = require('./geo');
+const { sendMessage } = require('./slack')
 require('dotenv').config();
 
 /* System notes:
@@ -14,40 +15,40 @@ require('dotenv').config();
 // Airtable
 const base = new Airtable({ apiKey: process.env.AIRTABLE_API_KEY }).base('appEwduS0FSdDmaan');
 
-const token = process.env.SLACK_XOXB;
-const channel = process.env.SLACK_CHANNEL_ID;
-const bot = new Slack({ token });
+// const token = process.env.SLACK_XOXB;
+// const channel = process.env.SLACK_CHANNEL_ID;
+// const bot = new Slack({ token });
 
-// This function actually sends the message to the slack channel
-const sendMessage = (errand, task, vols) => bot.chat
-  .postMessage({
-    token,
-    channel,
-    text: '',
-    blocks: [
-      {
-        type: 'section',
-        text: {
-          type: 'mrkdwn',
-          text: ':exclamation: *A new errand has been added!* :exclamation:',
-        },
-      },
-      errand,
-      task,
-      {
-        type: 'section',
-        text: {
-          type: 'mrkdwn',
-          text: ' ',
-        },
-      },
-    ],
-    attachments: [
-      {
-        blocks: vols,
-      },
-    ],
-  });
+// // This function actually sends the message to the slack channel
+// const sendMessage = (errand, task, vols) => bot.chat
+//   .postMessage({
+//     token,
+//     channel,
+//     text: '',
+//     blocks: [
+//       {
+//         type: 'section',
+//         text: {
+//           type: 'mrkdwn',
+//           text: ':exclamation: *A new errand has been added!* :exclamation:',
+//         },
+//       },
+//       errand,
+//       task,
+//       {
+//         type: 'section',
+//         text: {
+//           type: 'mrkdwn',
+//           text: ' ',
+//         },
+//       },
+//     ],
+//     attachments: [
+//       {
+//         blocks: vols,
+//       },
+//     ],
+//   });
 
 // Gets list of tasks from spreadsheet and adds to message text
 function formatTasks(row) {
